Accidental Mega-commit.
Moved all source files to a Python Package structured format to enable setuptools usage.
Added circleci config files and Makefile for automated testing using Docker Containers.
Added information to README on how to run test suite and install project with setuptools workflow.
Added skeleton requirements.txt
Moved example file to examples folder instead of in main package directory.
Removed cmt as a requirement as it is not set up properly for a simple pip installation, however I left in the Docker installation of cmt just in case it needs to be added back so it can still work with CI/Docker.